- integrates with calendar - improves authorization

+++ src/Http/Controllers/Tasks/Destroy.php
@@ -2,13 +2,18 @@
 
 namespace LaravelEnso\Tasks\Http\Controllers\Tasks;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use LaravelEnso\Tasks\Models\Task;
 use Illuminate\Routing\Controller;
 
 class Destroy extends Controller
 {
+    use AuthorizesRequests;
+
     public function __invoke(Task $task)
     {
+        $this->authorize('handle', $task);
+
         $task->delete();
 
         return [
